requirements.txt
torch
transformers
diffusers
segment-anything
git+https://github.com/openai/CLIP.git
pandas
Pillow
matplotlib
accelerate
